Extend error messages with cmd and output
